merge: 충돌 해결

+++ src/main/java/team05/integrated_feed_backend/module/post/repository/PostRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import team05.integrated_feed_backend.module.post.entity.Post;
 
 @Repository
-public interface PostRepository extends JpaRepository<Post, Integer>, PostCustomRepository {
+public interface PostRepository extends JpaRepository<Post, Long>, PostCustomRepository {
 }
